+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix localization of warning message when a ntoe has been reported. (thanks @L!)
 - Fixed contact list hydration bug where unfollows are not removed when follow counts do not change.
 
 ## [0.1 (101)] - 2023-12-15Z